(#944) Add help command
Add help command to display help when provided with choco help
