+++ src/chocolatey.tests/packages.config
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-
 <packages>
   <package id="log4net" version="2.0.3" targetFramework="net40" />
   <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net40" />
   <package id="Moq" version="4.2.1402.2112" targetFramework="net40" />
   <package id="NUnit" version="2.6.4" targetFramework="net40" />
   <package id="Should" version="1.1.12.0" />
+  <package id="SimpleInjector" version="2.5.0" targetFramework="net40" />
   <package id="TinySpec.NUnit" version="0.9.5" />
 </packages>
