+++ src/chocolatey/chocolatey.csproj
@@ -114,6 +114,7 @@
     </Compile>
     <Compile Include="AssemblyExtensions.cs" />
     <Compile Include="infrastructure.app\commands\ChocolateyInfoCommand.cs" />
+    <Compile Include="infrastructure.app\commands\ChocolateyHelpCommand.cs" />
     <Compile Include="infrastructure\cryptography\DefaultEncryptionUtility.cs" />
     <Compile Include="infrastructure\adapters\IEncryptionUtility.cs" />
     <Compile Include="infrastructure.app\validations\GlobalConfigurationValidation.cs" />
